Make breaches more ergonomic
Now breaches come wrapped in a collection class that responds to
`#to_hash`, making it simple to use them as a failure condition for the
interactor.

In fact, it was nice enough that I changed the default consequence to
use that behavior. Now the default consequence fails the interactor with
all of the breaches keys and the reason(s) they failed. See the readme
or the example in `BreachSet` for more information.