Update dry-validation dependency to 1.0
dry-validation has gone through many changes since this gem was
originally written. This change refactors the internals to ensure that
we can use a modern version of dry-validation with it.

There are no semantic changes in the way the gem works, but some of the
tests had to change to use the changes in the DSL within dry-validation.